coco3: separate low-level video methods into separate source file.
